requirements.txt
google-generativeai
flask
gunicorn
werkzeug